+++ frontend/components/Domain/Recipe/RecipeOrganizerSelector.vue
@@ -46,7 +46,7 @@ import { RecipeCategory, RecipeTag } from "~/lib/api/types/user";
 import { RecipeTool } from "~/lib/api/types/admin";
 import { useTagStore } from "~/composables/store/use-tag-store";
 import { useCategoryStore, useToolStore } from "~/composables/store";
-import { Organizer, RecipeOrganizer } from "~/types/recipe/organizers";
+import { Organizer, RecipeOrganizer } from "~/lib/api/types/non-generated";
 
 export default defineComponent({
   components: {
