+++ frontend/lib/api/types/non-generated.ts
@@ -23,3 +23,11 @@ export interface PaginationData<T> {
   total_pages: number;
   items: T[];
 }
+
+export type RecipeOrganizer = "categories" | "tags" | "tools";
+
+export enum Organizer {
+  Category = "categories",
+  Tag = "tags",
+  Tool = "tools",
+}
