--- frontend/types/recipe/organizers.ts
@@ -1,7 +1,0 @@
-export type RecipeOrganizer = "categories" | "tags" | "tools";
-
-export enum Organizer {
-  Category = "categories",
-  Tag = "tags",
-  Tool = "tools",
-}
